Ability to change data and amount that standard fee changes
We now have Client#current_rate which defines the current rate.  By setting Client#new_rate and Client#new_rate_from a new Price model is added to the client reflecting the new rate.

Validators have been added that check that both Client#new_rate and Client#new_rate_from are both set when we save the record.

Driven by this change, we have also restructured the UI code.  The ClientForm.svelte component now has two slot elements: `navigation-elements` that is a place to put buttons at the top of the form, and `form-elements` that is used to place extra fields into the form when required.

As we are now passing money back and forth we have also adopted using the Blueprints gem to generate the JSON in Rails more effectively.  Particular for passing currency, this is now converted into a JSON object that follows the same naming conventions as Dinero Javascript package.